Feature/dynamicvalidations

+++ app/Http/Controllers/GenericResourceController.php
@@ -32,6 +32,10 @@ class GenericResourceController extends Controller
             return $this->jsonError(['Model not found.'], 404);
         }
 
+        if ($this->validateInputsForResource($request->all(), $request->route('resource')) === false) {
+            return $this->jsonError(['Insufficient permissions.'], 403);
+        }
+
         return $this->jsonSuccess($model);
     }
 
@@ -41,7 +45,9 @@ class GenericResourceController extends Controller
      */
     public function store(Request $request)
     {
-        $this->validateInputsForResource($request->all(), $request->route('resource'));
+        if ($this->validateInputsForResource($request->all(), $request->route('resource')) === false) {
+            return $this->jsonError(['Insufficient permissions.'], 403);
+        }
 
         $model = GenericModel::create($request->all());
         if ($model->save()) {
@@ -62,7 +68,9 @@ class GenericResourceController extends Controller
             return $this->jsonError(['Model not found.'], 404);
         }
 
-        $this->validateInputsForResource($request->all(), $request->route('resource'));
+        if ($this->validateInputsForResource($request->all(), $request->route('resource')) === false) {
+            return $this->jsonError(['Insufficient permissions.'], 403);
+        }
 
         $model->fill($request->all());
         if ($model->save()) {
@@ -82,6 +90,10 @@ class GenericResourceController extends Controller
 
         if (!$model instanceof GenericModel) {
             return $this->jsonError(['Model not found.'], 404);
+        }
+
+        if ($this->validateInputsForResource($request->all(), $request->route('resource')) === false) {
+            return $this->jsonError(['Insufficient permissions.'], 403);
         }
 
         if ($model->delete()) {
